Moved out more enum functions
