+++ src/Engine/Objects/SpriteObject.h
@@ -3,7 +3,7 @@
 #include <vector>
 
 #include "Engine/Objects/Items.h"
-#include "Engine/Objects/SpriteObjectEnums.h"
+#include "Engine/Objects/SpriteEnums.h"
 #include "Engine/Objects/ActorEnums.h"
 #include "Engine/Spells/SpellEnums.h"
 #include "Engine/Pid.h"
